requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
tqdm>=4.65.0
pytest>=7.4.0
pytest-cov>=4.1.0